refactor: select torrent file list
